[lldb] Isolate reusable Darwin register contexts
