+++ lldb/source/Plugins/Process/Common/CMakeLists.txt
@@ -1,4 +1,10 @@
 add_lldb_library(lldbProcessContextSupport NO_PLUGIN_DEPENDENCIES
+  RegisterContext/Darwin/RegisterContextDarwin_arm.cpp
+  RegisterContext/Darwin/RegisterContextDarwin_arm64.cpp
+  RegisterContext/Darwin/RegisterContextDarwin_riscv32.cpp
+  RegisterContext/Darwin/RegisterContextDarwin_x86_64.cpp
+  RegisterContext/Darwin/RegisterContextMach_arm.cpp
+  RegisterContext/Darwin/RegisterContextMach_x86_64.cpp
   RegisterContext/POSIX/RegisterContextPOSIX_arm.cpp
   RegisterContext/POSIX/RegisterContextPOSIX_arm64.cpp
   RegisterContext/POSIX/RegisterContextPOSIX_loongarch64.cpp
