+++ lldb/source/Plugins/Process/Utility/CMakeLists.txt
@@ -14,14 +14,8 @@ add_lldb_library(lldbPluginProcessUtility
   NativeProcessSoftwareSingleStep.cpp
   NetBSDSignals.cpp
   OpenBSDSignals.cpp
-  RegisterContextDarwin_arm.cpp
-  RegisterContextDarwin_arm64.cpp
-  RegisterContextDarwin_riscv32.cpp
-  RegisterContextDarwin_x86_64.cpp
   RegisterContextDummy.cpp
   RegisterContextHistory.cpp
-  RegisterContextMach_arm.cpp
-  RegisterContextMach_x86_64.cpp
   RegisterContextMemory.cpp
   RegisterContextThreadMemory.cpp
   StopInfoMachException.cpp
