+++ llvm/utils/gn/secondary/lldb/source/Plugins/Process/Common/BUILD.gn
@@ -10,6 +10,12 @@ static_library("Common") {
 
   include_dirs = [ "//lldb/source" ]
   sources = [
+    "RegisterContext/Darwin/RegisterContextDarwin_arm.cpp",
+    "RegisterContext/Darwin/RegisterContextDarwin_arm64.cpp",
+    "RegisterContext/Darwin/RegisterContextDarwin_riscv32.cpp",
+    "RegisterContext/Darwin/RegisterContextDarwin_x86_64.cpp",
+    "RegisterContext/Darwin/RegisterContextMach_arm.cpp",
+    "RegisterContext/Darwin/RegisterContextMach_x86_64.cpp",
     "RegisterContext/POSIX/RegisterContextPOSIX_arm.cpp",
     "RegisterContext/POSIX/RegisterContextPOSIX_arm64.cpp",
     "RegisterContext/POSIX/RegisterContextPOSIX_loongarch64.cpp",
