+++ llvm/utils/gn/secondary/lldb/source/Plugins/Process/Utility/BUILD.gn
@@ -32,14 +32,8 @@ static_library("Utility") {
     "NativeProcessSoftwareSingleStep.cpp",
     "NetBSDSignals.cpp",
     "OpenBSDSignals.cpp",
-    "RegisterContextDarwin_arm.cpp",
-    "RegisterContextDarwin_arm64.cpp",
-    "RegisterContextDarwin_riscv32.cpp",
-    "RegisterContextDarwin_x86_64.cpp",
     "RegisterContextDummy.cpp",
     "RegisterContextHistory.cpp",
-    "RegisterContextMach_arm.cpp",
-    "RegisterContextMach_x86_64.cpp",
     "RegisterContextMemory.cpp",
     "RegisterContextThreadMemory.cpp",
     "StopInfoMachException.cpp",
